Updated Quiz, Model, and Answers
Updated this to get a randomly selected question and created Answers class to have answers object have a correct attribute which tells whether it's the right answer or not. The problem I am having right now with this is that for every frame after death a new quiz question number is generated, instead of staying as one/having the frames stop when you die.